added tests reworked astar for better heuristics
git-svn-id: svn://localhost/trunk@528 f2a0fb1d-5e3f-4af0-b388-fe59793aa704

+++ src/java/ironfist/astar/Node.java
@@ -14,13 +14,13 @@ public interface Node {
 
   Node[] getSuccessors();
 
-  void setG(int cost);
+  void setG(double cost);
 
-  int getG();
+  double getG();
 
-  void setH(int estimage);
+  void setH(double estimage);
 
-  int getH();
+  double getH();
 
-  int getF();
+  double getF();
 }
