+++ src/java/ironfist/astar/NodeFComparator.java
@@ -17,8 +17,8 @@ public class NodeFComparator implements Comparator {
    */
 
   public int compare(Object o1, Object o2) {
-    int f1 = ((Node) o1).getF();
-    int f2 = ((Node) o2).getF();
+    double f1 = ((Node) o1).getF();
+    double f2 = ((Node) o2).getF();
     return (f1 < f2 ? -1 : (f1 == f2 ? 0 : 1));
   }
